Fixed and optimised the code so that it runs on the Arduino Uno

+++ y2s1/ma2012-introduction-to-mechatronics-system-design/project/main-arduino/UltrasonicSensor.cpp
@@ -15,7 +15,7 @@ UltrasonicSensor::UltrasonicSensor(unsigned int trigger_pin,
 }
 
 // Function to get the distance in cm from the ultrasonic sensor
-float UltrasonicSensor::get_distance_in_cm() {
+unsigned int UltrasonicSensor::get_distance_in_cm() {
 
   // Reset the state of the ultrasonic sensor
   digitalWrite(this->_trigger_pin, LOW);
